refactor(job): migrate job logic to JobService and remove legacy repository functions
