+++ web-app/src/app/(app)/agents/[agentId]/jobs/layout.tsx
@@ -9,12 +9,12 @@ import {
 import DefaultLoading from "@/components/default-loading";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
 import { getAgentDescription, getAgentLegal, getAgentName } from "@/lib/db";
-import { retrieveJobsWithLimitedInformationByAgentId } from "@/lib/db/repositories";
 import {
   getAgentById,
   getAgentCreditsPrice,
   getAvailableAgentById,
   getFavoriteAgents,
+  JobService,
 } from "@/lib/services";
 
 import Footer from "./components/footer";
@@ -66,7 +66,10 @@ async function JobLayoutInner({ right, params, children }: JobLayoutProps) {
 
   const agentCreditsPrice = await getAgentCreditsPrice(agent);
   const favoriteAgents = await getFavoriteAgents();
-  const jobs = await retrieveJobsWithLimitedInformationByAgentId(agentId);
+  const jobs =
+    await JobService.getInstance().getJobsWithLimitedInformationByAgentId(
+      agentId,
+    );
   const availableAgent = await getAvailableAgentById(agentId);
 
   return (
