+++ web-app/src/lib/actions/job/action.ts
@@ -7,17 +7,12 @@ import { isJobError, JobErrorCode } from "@/lib/actions/types/error-codes/job";
 import { getSession } from "@/lib/auth/utils";
 import { JobWithStatus } from "@/lib/db";
 import {
-  retrieveJobByBlockchainIdentifier,
-  retrieveJobById,
-  updateJobNameById,
-} from "@/lib/db/repositories";
-import {
   jobDetailsNameFormSchema,
   JobDetailsNameFormSchemaType,
   startJobInputSchema,
   StartJobInputSchemaType,
 } from "@/lib/schemas";
-import { requestRefundJob, startJob } from "@/lib/services";
+import { JobService, requestRefundJob, startJob } from "@/lib/services";
 import { Err, Ok, Result } from "@/lib/ts-res";
 
 export async function startJobWithInputData(
@@ -185,7 +180,7 @@ export async function updateJobName(
     }
     const parsed = parsedResult.data;
 
-    const job = await retrieveJobById(jobId);
+    const job = await JobService.getInstance().getJobById(jobId);
     if (!job) {
       return Err({
         message: "Job not found",
@@ -202,7 +197,10 @@ export async function updateJobName(
     }
 
     // update job name
-    await updateJobNameById(jobId, parsed.name === "" ? null : parsed.name);
+    await JobService.getInstance().updateJobNameById(
+      jobId,
+      parsed.name === "" ? null : parsed.name,
+    );
     return Ok();
   } catch (error) {
     Sentry.withScope((scope) => {
@@ -245,7 +243,9 @@ export async function requestRefundJobByBlockchainIdentifier(
     }
 
     const foundJob =
-      await retrieveJobByBlockchainIdentifier(blockchainIdentifier);
+      await JobService.getInstance().getJobByBlockchainIdentifier(
+        blockchainIdentifier,
+      );
     if (!foundJob) {
       return Err({
         message: "Job not found",
