+++ web-app/src/lib/db/repositories/index.ts
@@ -1,3 +1,2 @@
 export * from "./invitation";
-export * from "./job";
 export { default as prisma } from "./prisma";
