+++ web-app/src/lib/services/index.ts
@@ -9,6 +9,7 @@ export { FiatTransactionService } from "./fiatTransaction.service";
 export * from "./invitation";
 export { InvitationService } from "./invitation.service";
 export * from "./job";
+export { JobService } from "./job.service";
 export { LockService } from "./lock.service";
 export { MemberService } from "./member.service";
 export { OrganizationService } from "./organization.service";
